--- Pulumi.pk8s.yaml
@@ -1,2 +1,0 @@
-config:
-  digitalocean-ts-k8s:domainName: do.jxp.io
