+++ cluster/Pulumi.yaml
@@ -1,3 +1,3 @@
-name: digitalocean-ts-k8s
+name: cluster
 runtime: nodejs
 description: Provision a DigitalOcean Kubernetes cluster and deploy to it
