+++ guestbook/Pulumi.yaml
@@ -1,0 +1,3 @@
+name: app
+runtime: nodejs
+description: Provision a DigitalOcean Kubernetes cluster and deploy to it
