+++ monitoring/index.ts
@@ -2,45 +2,15 @@
 
 import * as pulumi from "@pulumi/pulumi";
 import * as k8s from "@pulumi/kubernetes";
-import {kubeconfig} from "./k8s-cluster";
-import * as k8sService from "./k8s-service";
 import * as digitalocean from "@pulumi/digitalocean";
 
 
 const config = new pulumi.Config();
 const domainName = config.get("domainName");
 
-const provider = new k8s.Provider("do-k8s", {kubeconfig});
-
-const redisLeader = new k8sService.ServiceDeployment("redis-leader", {
-    image: "redis",
-    ports: [6379],
-  },
-  {provider: provider});
-
-const redisReplica = new k8sService.ServiceDeployment("redis-replica", {
-    image: "pulumi/guestbook-redis-replica",
-    ports: [6379],
-  },
-  {provider: provider});
-
-const frontend = new k8sService.ServiceDeployment("frontend", {
-    replicas: 3,
-    image: "pulumi/guestbook-php-redis",
-    ports: [80],
-    allocateIpAddress: true,
-    dnsName: "@"
-  },
-  {provider: provider});
-
-// outputs for guestbook frontend
-export const guestbookIp = frontend.ipAddress;
-export const guestbookUrl = pulumi.interpolate `http://${domainName}`;
+const provider = new k8s.Provider("do-k8s", { kubeconfig: process.env.KUBECONFIG });
 
 
-// ----------------------------------------------------------------------------
-// todo: break out the prometheus stack into its own component file
-// ----------------------------------------------------------------------------
 const prometheusNamespace = config.get("prometheusNamespace") || "prometheus";
 const appLabels = {
   app: "prometheus",
@@ -57,14 +27,13 @@ const prometheusNs = new k8s.core.v1.Namespace(prometheusNamespace, {
 });
 
 // Use Helm to install the Nginx prometheus controller
-// todo: figure out why recreate timed out when I changed this resource name
-const prometheusStack = new k8s.helm.v3.Release("prometheusstack", {
+const kube_pg_stack = new k8s.helm.v3.Release("kubepg", {
   chart: "kube-prometheus-stack",
   namespace: prometheusNamespace,
   repositoryOpts: {
     repo: "https://prometheus-community.github.io/helm-charts",
   },
-  // skipCrds: true,
+  skipCrds: false,
   values: {
     grafana: {
       // adminPassword: "admin",
@@ -77,20 +46,20 @@ const prometheusStack = new k8s.helm.v3.Release("prometheusstack", {
 }, {
   provider: provider,
   dependsOn: prometheusNs,
-  customTimeouts: { create: "10m" },
+  customTimeouts: { create: "20m" },
 });
 
 // Get the Grafana service so that we can extract its IP address
 const grafanaService: k8s.core.v1.Service = k8s.core.v1.Service.get(
   "grafana-service",
-  pulumi.interpolate `${prometheusNs.metadata.name}/${prometheusStack.status.name}-grafana`
+  pulumi.interpolate `${prometheusNs.metadata.name}/${kube_pg_stack.status.name}-grafana`
 )
 export const grafanaIngressIp = grafanaService.status.loadBalancer.ingress[0].ip
 
 // Get the Grafana secret created by the helm chart so that we can extract the default admin creds
 const grafanaSecret: k8s.core.v1.Secret = k8s.core.v1.Secret.get(
   "grafana-secret",
-  pulumi.interpolate `${prometheusNs.metadata.name}/${prometheusStack.status.name}-grafana`
+  pulumi.interpolate `${prometheusNs.metadata.name}/${kube_pg_stack.status.name}-grafana`
 )
 
 export const grafanaAdminUser = grafanaSecret.data["admin-user"].apply((value) => {
@@ -102,7 +71,6 @@ export const grafanaAdminPass = grafanaSecret.data["admin-password"].apply((valu
   const buffer = Buffer.from(value, "base64");
   return buffer.toString("utf-8");
 });
-
 
 if (domainName) {
   const aRecord = new digitalocean.DnsRecord("do-grafana-a-rec", {
